requirements.txt
flask==3.0.0
numpy==2.1.3
opencv-python-headless==4.8.1.78
pillow==11.0.0
pytesseract==0.3.10
pandas==2.2.3
scikit-learn==1.5.2
spacy==3.7.2
transformers==4.36.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2+cpu
langdetect==1.0.9
langid==1.1.6
